Implement MapLatencyRecorder
